+++ README.md
@@ -301,6 +301,7 @@ The following resources are available in the InSpec GCP Profile
 | [google_dlp_job_trigger](docs/resources/google_dlp_job_trigger.md)                                   | [google_dlp_job_triggers](docs/resources/google_dlp_job_triggers.md)                         |
 | [google_dlp_job](docs/resources/google_dlp_job.md)                                                                    | [google_dlp_jobs](docs/resources/google_dlp_jobs.md)                                 |
 | [google_dataproc_metastore_service](docs/resources/google_dataproc_metastore_service.md)                                                                    | [google_dataproc_metastore_services](docs/resources/google_dataproc_metastore_services.md)                                 |
+| [google_dataproc_session](docs/resources/google_dataproc_session.md)                                                                    | [google_dataproc_sessions](docs/resources/google_dataproc_sessions.md)                                 |
 | [google_dlp_stored_info_type](docs/resources/google_dlp_stored_info_types.md)                         | [google_dlp_stored_info_types](docs/resources/google_dlp_stored_info_types.md)                         |
 | [google_filestore_instance](docs/resources/google_filestore_instance.md)                                             | [google_filestore_instances](docs/resources/google_filestore_instances.md)                                                     |
 | [google_iam_organization_custom_role](docs/resources/google_iam_organization_custom_role.md)                         | [google_iam_organization_custom_roles](docs/resources/google_iam_organization_custom_roles.md)                                 |
