+++ libraries/google/dataproc/property/session_jupyter_session.rb
@@ -16,38 +16,20 @@
 module GoogleInSpec
   module Dataproc
     module Property
-      class ProjectLocationSessionEnvironmentConfigExecutionConfig
-        attr_reader :service_account
+      class SessionJupyterSession
+        attr_reader :kernel
 
-        attr_reader :network_uri
-
-        attr_reader :subnetwork_uri
-
-        attr_reader :network_tags
-
-        attr_reader :kms_key
-
-        attr_reader :idle_ttl
-
-        attr_reader :ttl
-
-        attr_reader :staging_bucket
+        attr_reader :display_name
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @service_account = args['serviceAccount']
-          @network_uri = args['networkUri']
-          @subnetwork_uri = args['subnetworkUri']
-          @network_tags = args['networkTags']
-          @kms_key = args['kmsKey']
-          @idle_ttl = args['idleTtl']
-          @ttl = args['ttl']
-          @staging_bucket = args['stagingBucket']
+          @kernel = args['kernel']
+          @display_name = args['displayName']
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationSessionEnvironmentConfigExecutionConfig"
+          "#{@parent_identifier} SessionJupyterSession"
         end
       end
     end
