+++ libraries/google/dataproc/property/session_labels.rb
@@ -16,35 +16,17 @@
 module GoogleInSpec
   module Dataproc
     module Property
-      class ProjectLocationSessionRuntimeInfoCurrentUsage
-        attr_reader :milli_dcu
-
-        attr_reader :shuffle_storage_gb
-
-        attr_reader :milli_dcu_premium
-
-        attr_reader :shuffle_storage_gb_premium
-
-        attr_reader :milli_accelerator
-
-        attr_reader :accelerator_type
-
-        attr_reader :snapshot_time
+      class SessionLabels
+        attr_reader :additional_properties
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @milli_dcu = args['milliDcu']
-          @shuffle_storage_gb = args['shuffleStorageGb']
-          @milli_dcu_premium = args['milliDcuPremium']
-          @shuffle_storage_gb_premium = args['shuffleStorageGbPremium']
-          @milli_accelerator = args['milliAccelerator']
-          @accelerator_type = args['acceleratorType']
-          @snapshot_time = args['snapshotTime']
+          @additional_properties = args['additionalProperties']
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationSessionRuntimeInfoCurrentUsage"
+          "#{@parent_identifier} SessionLabels"
         end
       end
     end
