+++ libraries/google/dataproc/property/session_runtime_config_repository_config_pypi_repository_config.rb
@@ -13,21 +13,20 @@
 #     CONTRIBUTING.md located at the root of this package.
 #
 # ----------------------------------------------------------------------------
-require 'google/dataproc/property/projectlocationsession_runtime_config_repository_config_pypi_repository_config'
 module GoogleInSpec
   module Dataproc
     module Property
-      class ProjectLocationSessionRuntimeConfigRepositoryConfig
-        attr_reader :pypi_repository_config
+      class SessionRuntimeConfigRepositoryConfigPypiRepositoryConfig
+        attr_reader :pypi_repository
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @pypi_repository_config = GoogleInSpec::Dataproc::Property::ProjectLocationSessionRuntimeConfigRepositoryConfigPypiRepositoryConfig.new(args['pypiRepositoryConfig'], to_s)
+          @pypi_repository = args['pypiRepository']
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationSessionRuntimeConfigRepositoryConfig"
+          "#{@parent_identifier} SessionRuntimeConfigRepositoryConfigPypiRepositoryConfig"
         end
       end
     end
