+++ libraries/google/dataproc/property/session_runtime_info_current_usage.rb
@@ -16,17 +16,35 @@
 module GoogleInSpec
   module Dataproc
     module Property
-      class ProjectLocationSessionRuntimeConfigRepositoryConfigPypiRepositoryConfig
-        attr_reader :pypi_repository
+      class SessionRuntimeInfoCurrentUsage
+        attr_reader :milli_dcu
+
+        attr_reader :shuffle_storage_gb
+
+        attr_reader :milli_dcu_premium
+
+        attr_reader :shuffle_storage_gb_premium
+
+        attr_reader :milli_accelerator
+
+        attr_reader :accelerator_type
+
+        attr_reader :snapshot_time
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @pypi_repository = args['pypiRepository']
+          @milli_dcu = args['milliDcu']
+          @shuffle_storage_gb = args['shuffleStorageGb']
+          @milli_dcu_premium = args['milliDcuPremium']
+          @shuffle_storage_gb_premium = args['shuffleStorageGbPremium']
+          @milli_accelerator = args['milliAccelerator']
+          @accelerator_type = args['acceleratorType']
+          @snapshot_time = args['snapshotTime']
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationSessionRuntimeConfigRepositoryConfigPypiRepositoryConfig"
+          "#{@parent_identifier} SessionRuntimeInfoCurrentUsage"
         end
       end
     end
