--- test/integration/verify/controls/google_dataproc_project_location_session.rb
@@ -1,52 +1,0 @@
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-
-title 'Test GCP google_dataproc_project_location_session resource.'
-
-gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
-
-  project_location_session = input('project_location_session', value: {
-  "name": "value_name",
-  "parent": "value_parent",
-  "uuid": "value_uuid",
-  "create_time": "value_createtime",
-  "state": "value_state",
-  "state_message": "value_statemessage",
-  "state_time": "value_statetime",
-  "creator": "value_creator",
-  "user": "value_user",
-  "session_template": "value_sessiontemplate"
-}, description: 'project_location_session description')
-control 'google_dataproc_project_location_session-1.0' do
-  impact 1.0
-  title 'google_dataproc_project_location_session resource test'
-
-  describe google_dataproc_project_location_session(name: project_location_session['name']) do
-  	it { should exist }
-  	its('name') { should cmp project_location_session['name'] }
-  	its('uuid') { should cmp project_location_session['uuid'] }
-  	its('create_time') { should cmp project_location_session['create_time'] }
-  	its('state') { should cmp project_location_session['state'] }
-  	its('state_message') { should cmp project_location_session['state_message'] }
-  	its('state_time') { should cmp project_location_session['state_time'] }
-  	its('creator') { should cmp project_location_session['creator'] }
-  	its('user') { should cmp project_location_session['user'] }
-  	its('session_template') { should cmp project_location_session['session_template'] }
-
-  end
-
-  describe google_dataproc_project_location_session(name: "does_not_exit") do
-  	it { should_not exist }
-  end
-end
