--- test/integration/verify/controls/google_dataproc_project_location_sessions.rb
@@ -1,38 +1,0 @@
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-
-title 'Test GCP google_dataproc_project_location_sessions resource.'
-
-gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
-
-  project_location_session = input('project_location_session', value: {
-  "name": "value_name",
-  "parent": "value_parent",
-  "uuid": "value_uuid",
-  "create_time": "value_createtime",
-  "state": "value_state",
-  "state_message": "value_statemessage",
-  "state_time": "value_statetime",
-  "creator": "value_creator",
-  "user": "value_user",
-  "session_template": "value_sessiontemplate"
-}, description: 'project_location_session description')
-control 'google_dataproc_project_location_sessions-1.0' do
-  impact 1.0
-  title 'google_dataproc_project_location_sessions resource test'
-
-      describe google_dataproc_project_location_sessions(parent: project_location_session['parent']) do
-      it { should exist }
-    end
-end
