+++ test/integration/verify/controls/google_dataproc_sessions.rb
@@ -1,0 +1,43 @@
+# ----------------------------------------------------------------------------
+#
+#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
+#
+# ----------------------------------------------------------------------------
+#
+#     This file is automatically generated by Magic Modules and manual
+#     changes will be clobbered when the file is regenerated.
+#
+#     Please read more about how to change this file in README.md and
+#     CONTRIBUTING.md located at the root of this package.
+#
+# ----------------------------------------------------------------------------
+
+title 'Test GCP google_dataproc_sessions resource.'
+
+gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
+
+session = input('session', value: {
+  "name": "projects/ppradhan/locations/us-central1/sessions/inspec-runtime-0ae1-b98f4174",
+  "parent": "projects/ppradhan/locations/us-central1",
+  "uuid": "0b8d8483-b1ff-4f67-8e34-8d57987430a2",
+  "create_time": "2024-08-19T06:59:23.721658Z",
+  "state": "value_state",
+  "state_message": "value_statemessage",
+  "state_time": "value_statetime",
+  "creator": "balasubramanian.s@chef.io",
+  "user": "value_user",
+  "session_template": "projects/ppradhan/locations/us-central1/sessionTemplates/inspec-runtime-0ae1"
+}, description: 'session description')
+control 'google_dataproc_sessions-1.0' do
+  impact 1.0
+  title 'google_dataproc_sessions resource test'
+
+    describe google_dataproc_sessions(parent: session['parent']) do
+      it { should exist }
+      its('names') { should include session['name'] }
+      its('uuids') { should include session['uuid'] }
+      its('create_times') { should include session['create_time'] }
+      its('creators') { should include session['creator'] }
+      its('session_templates') { should include session['session_template'] }
+    end
+end
